Break out client functions on backend
